Style country images and navbar
