+++ src/styling/Navbar.css
@@ -1,3 +1,6 @@
+nav {
+  background-color: #4369B2;
+}
 .back {
   border: none;
   background: transparent;
